Spark 的WordCount 成功，并解决了本地运行问题

+++ java-wordcount/src/main/java/com/spark/java/spark/TestSpark.java
@@ -56,10 +56,10 @@ public class TestSpark {
         });
         String outFile = "/root/zhanglu/data/clean_data";
         File file = new File(outFile);
-        if(file.exists()){
+        if(file.exists() && file.isDirectory()){
             file.delete();
         }
-        output.saveAsTextFile("/root/zhanglu/data/clean_data");
+        output.saveAsTextFile(outFile);
         context.close();
     }
 }
